Refactor GatorGradleConfig to better split commands into shellwords
Previously, using a quote (") in a command string could cause some
confusing results, and escaping was not possible. Now, both ' and " can
be used to delineate shell words, and either can contain the other.
Escaping quotes is still not possible, but using alternating quote types
should provide the same abilities.